Refactored footer into components, removed min-width usage in media queries https://github.com/SoftUni-Internal/exam-systems-issues/issues/367
