+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/contests/contest-tasks-navigation/ContestTasksNavigation.module.scss
@@ -14,6 +14,10 @@
   padding: 35px 0 5px;
   position: relative;
   width: 100%;
+
+  @media only screen and (max-width: variables.$r-large-desktop-max) {
+    font-size: variables.$f-size-16;
+  }
 }
 
 .tasksListSideNavigation {
@@ -35,12 +39,6 @@
 
     &.selected {
       color: colors.$color-btn-secondary;
-    }
-
-    .taskListItem {
-      @media only screen and (max-width: variables.$r-large-desktop-max) {
-        font-size: variables.$f-size-16;
-      }
     }
   }
 }
