+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/problems/problem-submissions/ProblemSubmissions.module.scss
@@ -4,6 +4,14 @@
 .submissionResultsContent {
   display: flex;
   flex-direction: column;
+
+  @media only screen and (max-width: variables.$r-large-desktop-max) {
+    padding: 10px;
+
+    .submissionsList > ul {
+      font-size: variables.$f-size-16;
+    }
+  }
 }
 
 .refreshBtn {
@@ -13,15 +21,5 @@
   &:hover {
     background: none;
     cursor: pointer;
-  }
-
-  @media only screen and (max-width: variables.$r-large-desktop-max) {
-    .submissionResultsContent {
-      padding: 10px;
-    }
-
-    .submissionsList > ul {
-      font-size: variables.$f-size-16;
-    }
   }
 }
