+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/layout/footer/FooterListSection.module.scss
@@ -1,0 +1,33 @@
+@use 'src/styles/colors';
+@use 'src/styles/fonts';
+@use 'src/styles/components';
+@use 'src/styles/variables';
+
+.footerListSection,
+%footerListSection {
+  align-content: start;
+  color: colors.$text-light-gray-color;
+  display: flex;
+  flex-flow: column;
+
+  .linksList {
+    align-items: flex-start;
+  }
+
+  .sectionHeading {
+    color: colors.$white-color;
+    font-size: variables.$f-size-18;
+    padding-left: 10px;
+    padding-top: 0;
+  }
+
+  .link {
+    @extend %whiteOnHover;
+
+    color: colors.$text-light-gray-color;
+    font-size: variables.$f-size-16;
+    margin-top: 10px;
+    padding-left: 10px;
+    text-align: start;
+  }
+}
