+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/layout/footer/FooterNavigation.tsx
@@ -1,0 +1,92 @@
+import React, { useCallback, useMemo } from 'react';
+
+import { LinkButton, LinkButtonType } from '../../components/guidelines/buttons/Button';
+import GithubIcon from '../../components/guidelines/icons/GitHubIcon';
+import YouTubeIcon from '../../components/guidelines/icons/YouTubeIcon';
+
+import FooterListSection from './FooterListSection';
+import logo from './softuni-logo-horizontal-white.svg';
+
+import styles from './FooterNavigation.module.scss';
+
+interface IFooterLinkType {
+    text: string;
+    url: string;
+}
+
+const FooterNavigation = () => {
+    const learnLinks = useMemo(() => Array.from<IFooterLinkType>([
+        {
+            text: 'Professional Programs',
+            url: 'https://learn.softuni.org/catalog#program',
+        },
+        {
+            text: 'Courses',
+            url: 'https://learn.softuni.org/catalog#opencourse',
+        },
+        {
+            text: 'Open Lessons',
+            url: 'https://learn.softuni.org/catalog#openlesson',
+        } ]), []);
+
+    const aboutLinks = useMemo(() => Array.from<IFooterLinkType>([
+        {
+            text: 'Contact us',
+            url: 'https://learn.softuni.org/contact',
+        },
+        {
+            text: 'Careers',
+            url: 'https://learn.softuni.org/careers',
+        },
+        {
+            text: 'Blog',
+            url: 'https://softuni.org/',
+        } ]), []);
+
+    const renderSystemInfoAndLinksSection = useCallback(() => (
+        <div className={styles.systemInfoAndLinksContainer}>
+            <span className={styles.systemInfo}>
+                © 2011-2023 - Open Judge System (OJS)
+            </span>
+            <span className={styles.links}>
+                <LinkButton
+                  to="https://github.com/NikolayIT/OpenJudgeSystem"
+                  type={LinkButtonType.plain}
+                  isToExternal
+                >
+                    <GithubIcon />
+                </LinkButton>
+                <LinkButton
+                  to="https://www.youtube.com/channel/UCqvOk8tYzfRS-eDy4vs3UyA"
+                  type={LinkButtonType.plain}
+                  isToExternal
+                >
+                    <YouTubeIcon />
+                </LinkButton>
+            </span>
+        </div>
+    ), []);
+
+    const renderLogo = useCallback(() => (
+        <div className={styles.footerLogo}>
+            <img src={logo} alt="softuni logo" />
+        </div>
+    ), []);
+
+    return (
+        <div className={styles.content}>
+            {renderLogo()}
+            <FooterListSection
+              links={learnLinks}
+              sectionHeadingText="Learn"
+            />
+            <FooterListSection
+              links={aboutLinks}
+              sectionHeadingText="About"
+            />
+            {renderSystemInfoAndLinksSection()}
+        </div>
+    );
+};
+
+export default FooterNavigation;
