+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/layout/footer/PageFooter.module.scss
@@ -10,77 +10,3 @@
   color: colors.$text-light-gray-color;
   min-height: 30vh;
 }
-
-.content {
-  @extend %contentSize;
-
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  margin: 0 auto;
-  padding: 60px 0;
-}
-
-.footerLearnLinksWrapper,
-%footerLearnLinksWrapper {
-  align-content: flex-start;
-  color: colors.$text-light-gray-color;
-  display: flex;
-  flex-flow: column;
-  justify-content: flex-start;
-
-  .sectionHeading {
-    color: colors.$white-color;
-    font-size: variables.$f-size-18;
-    padding-left: 10px;
-    padding-top: 0;
-  }
-
-  .link {
-    @extend %whiteOnHover;
-
-    color: colors.$text-light-gray-color;
-    font-size: variables.$f-size-16;
-    margin-top: 10px;
-    padding-left: 10px;
-    text-align: start;
-  }
-}
-
-.footerLogo {
-  max-width: 7vw;
-}
-
-.footerAboutLinksWrapper {
-  @extend %footerLearnLinksWrapper;
-}
-
-.systemInfoAndLinksContainer {
-  align-content: flex-start;
-  align-items: flex-start;
-  display: flex;
-  flex-direction: row;
-  gap: 30px;
-  justify-content: flex-start;
-}
-
-.systemInfo {
-  // body
-}
-
-.links {
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: row;
-  gap: 15px;
-
-  svg {
-    color: colors.$text-light-gray-color;
-    height: 1.2vw;
-    width: 1.2vw;
-
-    &:hover {
-      color: colors.$white-color;
-    }
-  }
-}
